Lab 1
You need to write a function analysis that parses and counts the learning data in the ~/Code/analysis.py Python file. 
The analysis function takes two arguments. The first parameter is the learning data json file name, 
the file content format is consistent with the user_study.json file format, and the second parameter is the user ID. 
You need to parse the json file in the function, and count the number of learning times and the total number of minutes 
of the user ID specified by the second parameter, that is, the function will return two values, the first one is the learning
time of the specified user. The second is the total number of minutes of learning for the specified user.